pubsub_app/requirements.txt
google-cloud-pubsub
google-cloud-bigquery
gunicorn